net: process_event() submits chunked events.
In block_process_event(), prof_dev may be closed. Causes `struct tcp_client_socket`
and `struct event_block` to be released, and perf-prof got SIGSEGV.

net submits chunked events. Able to process a batch of events when obtaining a
reference to prof_dev.

Simplifying the tcp_ref() code, getting the reference ensures that the tcp socket is
not closed, and the tcp_client_socket object is not released. When hanging up,
ops->disconnect() ensures that tcp_close() is called to release the reference.

Signed-off-by: duanery <corcpp@foxmail.com>