order: heap sort: use heapclock as heap sort timestamp
The heapclock serves as a common timestamp for heap sorting. Events may come from
perf_mmap and stream (such as tcp, cdev, etc). The time conversion from different
sources may be inconsistent, so they need to be uniformly converted to heapclock
before they can be sorted together.

If there are no stream events, heap sorting can be done using only perfclock, so
heapclock is perfclock.

If there are stream events, the push side and pull side need to use consistent time
conversion. In this scenario, heapclock is evclock.

The `lost_start' and `lost_end' of prof->lost() need to pass evclock, so convert the
heapclock to evclock is needed. And additionally provides heapclock to perfclock
conversion.

Fixes: 26af72f ("order: heap sort: support timestamp conversion")
Signed-off-by: duanery <corcpp@foxmail.com>