order: heap sort: All order-enabled devices are sorted together
Sorting together only adjusts the order of event processing and does not affect the
forwarding and close of the device.

There are 2 categories: forwarding source, cloned device.

The forwarding source and target are originally sorted together as a whole. All on
the `heap_event_list' linked list.

The cloned device (for example, the --ptrace option) should be a whole, but because
the isolation of the device cannot be achieved, it is also an improvement to be
able to perform heap sorting together.

These 2 categories can be combined.

  perf-prof trace -e x:yy,task-state --ptrace --order -- multi_thread_app
  x:yy and task-state events will be sorted together and output.

This does not affect the close of the cloned device.

Signed-off-by: duanery <corcpp@foxmail.com>