Get the symbols of the process under the non-init mount namespace
When we need to read symbols of non-default mntns, enter the corresponding
mount namespace through setns(), and return to the default mount namespace
after reading the symbols.

For container scenarios, we can avoid parsing the root directory from
/proc/<pid>/mounts and overlayfs.

For object, it is determined by mntns and name, and a reference to mntns is
obtained.

Signed-off-by: duanery <corcpp@foxmail.com>