tep: For kprobe and uprobe only sampling events require IDs
For kprobe and uprobe events, they are similar to tracepoint, but common_type=0, so
PERF_SAMPLE_ID must be used to distinguish events between tracepoint and kprobe and
uprobe events.

For sampling events, multiple different events may be written to a same ringbuffer,
and it is impossible to distinguish which event it is during processing.

For non-sampling events, you can use read(fd) to obtain the count. This fd can
distinguish different events.

Signed-off-by: duanery <corcpp@foxmail.com>