Demangle c++, java, rust symbol names
By default, cplus-demangle is used, which is defined in libiberty and requires the
binutils-devel package. Secondly, cxa-demangle is used, which is in libstdc++.

Define `HAVE_CPLUS_DEMANGLE_SUPPORT` and `HAVE_CXA_DEMANGLE_SUPPORT` macros for .c
files. libbfd is not used, as its static library is too large.

Generally speaking, the length of the demangled string will be smaller than the
original one, and not all symbols are used. Therefore, use `obj->strs_cap` as the
maximum buffer length for demangled string. If the buffer is full, no more demangles
are performed.

Signed-off-by: duanery <corcpp@foxmail.com>