breakpoint: Decode the MOV instruction to get the written value
Tracking memory writes through breakpoints, we can actually obtain the written
values. Different written values can be more helpful for analysis.

For value that will not change, you can directly read the kernel memory address
using kcore_read(). However, when perf-prof processes the sampling event, it is
already too late and it is difficult to ensure that the value does not change.

We can read the instruction and decode the memory address(destination operand)
to be written and the value(source operand) to be written. Since instructions
rarely change and all general registers are sampled, they can be fully decoded.

This patch only supports decoding MOV instructions, and the destination operand
is a memory address, and determines whether the memory address is a breakpoint
address. Because the data write breakpoint is a trap-class exception, breakpoint
will reported immediately following the execution of the trapping instruction. We
need to go back to find the instruction boundary. At the same time, the MOV
instruction that writes to memory does not change any general registers. Finally,
we can safely decode the source operand.

Signed-off-by: duanery <corcpp@foxmail.com>