breakpoint: Decode the ADD instruction to get the written value
This patch supports decoding ADD instructions, and the destination operand is a
memory address, and determines whether the memory address is a breakpoint address.

The ADD instruction must know the original value in memory, otherwise it is unsafe.
For the MOV instruction, it is always safe to get the original value. Others need
to be based on MOV. It is also safe to calculate the written value. And, we need
to ensure the order of writes.

Safe means that the written values we obtain are correct and accurate. Also, if an
unsupported instruction writes, it is unsafe.

We have sampled all sequential writes to a given memory address, and we have also
obtained a sequence of write instructions, so we can safely get all the written
values.

Continue: f9eac0f ("breakpoint: Decode the MOV instruction to get the written value")
Signed-off-by: duanery <corcpp@foxmail.com>