breakpoint: Decode the CMPXCHG instruction
CMPXCHG is a two-byte instruction. And the written value can be obtained by using
only regs. So it is safe.

Operation
TEMP := DEST
IF accumulator = TEMP
    THEN
        ZF := 1;
        DEST := SRC;
    ELSE
        ZF := 0;
        accumulator := TEMP;
        DEST := TEMP;
FI;

Check the ZF flag. If ZF=1, the written value is equal to the source operand.
Otherwise, the written value is equal to AL/rAX.

CMPXCHG is a key instruction for kernel atomic operations, spin locks, etc.
This support is more valuable.

Signed-off-by: duanery <corcpp@foxmail.com>