breakpoint: Decode the BTS/BTR/BTC instruction
For these Bit Test instructions, if BitBase is a memory address, the BitOffset
has different ranges depending on the operand size.

Operand Size  Immediate BitOffset  Register BitOffset
16            0 to 15              -2^15 to 2^15-1
32            0 to 31              -2^31 to 2^31-1
64            0 to 63              -2^63 to 2^63-1

The actual memory address accessed is BitBase + BitOffset/8. We need to align it to
the operand size boundary to ensure that the destination operand address overlaps
with the breakpoint range.

Signed-off-by: duanery <corcpp@foxmail.com>