bpf:kvm-exit: Collect unified run_delay for vcpu
sched_latency is the latency of a process switching out and back in. For non-HLT
vmexit, this is basically equal to run_delay. For HLT vmexit, it includes run_delay,
but is more of the vcpu sleep time. This sched_latency is not easy to understand.

For bpf-prog:
We use the kernel's sched_info.run_delay data to get the vcpu's run_delay. We use
the fact that the sched_switch event happens before the update of sched_info.run_delay.
So, we read the old value in sched_switch and the new value in kvm_entry to calculate
run_delay.

For perf-prof:
1. We accumulate run_delay for different exit_reasons. This gives the ratio of run_delay
   to total exit latency.
2. For output2, the max-latency, for HLT, we use run_delay, and for others, we use vmexit
   latency. For HLT, run_delay is the real delay, which can help us analyze.
3. --than, Same as 2.

Signed-off-by: duanery <corcpp@foxmail.com>