order: heap sort: Allow skipping of lost detection
When the device does not care about the correctness of lost, such as if profiler::lost
is not provided, heap sorting does not perform lost detection.

This is because lost detection may cause a large pause, completely blocking the heap sort.
For example:
  # kill -USR2 $(pidof perf-prof)
  order: unordered 0 fixed 0
  order: lost 9050 maybe 19306 pause 281 pause_time 121153222884256 <== 121153 seconds

Furthermore, we need to control the maximum pause time.

Signed-off-by: duanery <corcpp@foxmail.com>