split-lock: counts the RIP where lock occurs
The lock instruction may occur in the guest or the host, in the user or in the
kernel. The perf_event_header.misc field can be used to distinguish these
situations. For the lock instruction in the guest, we need to identify the
corresponding process within the VM and fix the code alignment to cache line
boundaries.

TEST CASE
$ perf-prof split-lock -T 1000 -i 1000 --test
2025-10-17 18:43:00.916150 split-lock
      SAMPLED_RIP     N
 000000000042beb5   369 host-user
 CPU  SPLIT_LOCKS  RUN%
   1       201257   100
   2       168183   100

Signed-off-by: duanery <corcpp@foxmail.com>